--- demo/src/test/java/com/github/tkaczenko/auditor/demo/OutboundFeignIntegrationTests.java
@@ -1,3 +1,0 @@
-package com.github.tkaczenko.auditor.demo;
-
-public class OutboundFeignIntegrationTests {}
