+++ demo/src/test/java/com/github/tkaczenko/auditor/demo/inbound/InboundIntegrationTests.java
@@ -1,10 +1,10 @@
-package com.github.tkaczenko.auditor.demo;
+package com.github.tkaczenko.auditor.demo.inbound;
 
 import static com.github.tkaczenko.auditor.demo.util.FileUtils.readSystemResource;
 import static com.github.tkaczenko.auditor.demo.util.Files.Inbound;
 
-import com.github.tkaczenko.auditor.demo.common.AbstractIntegrationTest;
-import com.github.tkaczenko.auditor.demo.common.IntegrationTestScenario;
+import com.github.tkaczenko.auditor.demo.AbstractIntegrationTest;
+import com.github.tkaczenko.auditor.demo.IntegrationTestScenario;
 import com.github.tkaczenko.auditor.demo.model.dto.ErrorResponse;
 import com.github.tkaczenko.auditor.demo.model.dto.Response;
 import io.restassured.response.ValidatableResponse;
