+++ demo/src/test/java/com/github/tkaczenko/auditor/demo/outbound/feign/OutboundFeignIntegrationTests.java
@@ -1,0 +1,3 @@
+package com.github.tkaczenko.auditor.demo.outbound.feign;
+
+public class OutboundFeignIntegrationTests {}
